requirements.txt
anyascii==0.3.3
asgiref==3.10.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.7
django-filter==25.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.7
django-taggit==6.1.0
django-tasks==0.8.1
django-treebeard==4.7.1
djangorestframework==3.16.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filelock==3.20.0
filetype==1.2.0
idna==3.11
laces==0.1.2
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
pillow==11.3.0
pillow_heif==1.1.1
platformdirs==4.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
wagtail==7.1.2
Willow==1.12.0
plotly==6.3.1
python-dotenv==1.1.1
python-docx==1.2.0
python-pptx==1.0.2
pypdf==6.7.0
nltk==3.9.2
whitenoise==6.11.0
